+++ clients/include/testing_syr_batched.hpp
@@ -28,7 +28,7 @@ hipblasStatus_t testing_syr_batched(const Arguments& argus)
     int               batch_count = argus.batch_count;
 
     int    abs_incx = incx < 0 ? -incx : incx;
-    size_t A_size   = lda * N;
+    size_t A_size   = size_t(lda) * N;
 
     T h_alpha = argus.get_alpha<T>();
